requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Werkzeug==3.0.1
redis==5.0.1
flask-caching==2.1.0
flask-compress==1.14
gunicorn==21.2.0
Flask-WTF==1.2.1
WTForms==3.1.1
Flask-Limiter==3.5.0
bleach==6.1.0
python-dotenv==1.0.0

# Dependências de teste (opcional - instalar com requirements-test.txt)
# pytest==7.4.3
# pytest-flask==1.3.0
# pytest-cov==4.1.0
# factory-boy==3.3.0
# faker==20.1.0
# playwright==1.40.0

# Dependências para documentação da API
flask-restx==1.3.0
flask-cors==4.0.0
marshmallow==3.20.1
apispec==6.3.0
apispec-webframeworks==0.5.2